reduce merge conflicts by restoring legacy test paths
