+++ spring_aop_xml/src/main/java/org/spring/test/TestSpring.java
@@ -1,6 +1,5 @@
 package org.spring.test;
 
-import org.junit.jupiter.api.Test;
 import org.spring.action.UserAction;
 import org.spring.model.User;
 import org.springframework.beans.factory.BeanFactory;
@@ -10,8 +9,7 @@ public class TestSpring {
 	//创建spring工厂
 	private BeanFactory factory = new ClassPathXmlApplicationContext("applicationContext.xml");
 	
-	@Test
-	public void testAAUser() {
+		public void testAXUser() {
 		UserAction ua = factory.getBean("userAction",UserAction.class);
 		User u = new User(1,"张三",2000,22);
 		ua.setUser(u);
